pyproject.toml
[tool.poetry]
name = "routine"
version = "0.1.0"
description = "An unofficial Python API for the Routine planner app."
authors = ["Florian Schäfer <florian.joh.schaefer@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pydantic = "^1.10.8"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
black = "^23.3.0"
ruff = "^0.0.270"
mypy = "^1.3.0"
types-requests = "^2.31.0.1"

[tool.black]
line-length = 119
target-version = ['py311']

[tool.ruff]
ignore = [
    "E501", # line length violations are handled by black
    "D100", "D104", "D107"  # Do no require docstring for public modules/packages
]
select = [
    # "D",  # pydocstyle
    "E",  # pycodestyle (Error)
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W"  # pycodestyle (Warning)
]
pydocstyle.convention = "google"
line-length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"